+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 #### 2018.05.04
 - 关联汉字的五行属性，来源[便民吧](http://www.bm8.com.cn/wuxing/%AD%8F.htm)
 #### 2018.05.05
+- bugfix [GB2312](https://www.ibm.com/developerworks/cn/java/j-lo-chinesecoding/index.html)只包含6000+汉字，7000常用汉字中部分不在编码内，需要使用GBK。
+- **#TO DO#** spring boot的单测